requirements.txt
accelerate
bitsandbytes
diffusers
discord.py
GitPython
numpy
opencv-python
peft
protobuf
psd-tools
PyGithub
requests
sentencepiece
transformers
yt-dlp
